# openapi.yaml
openapi: 3.1.0
info:
  title: Smart Home Symcon Service
  description: API сервис управления устройствами умного дома.
  contact:
    name: Alexey Kuznetsov
    url: https://github.com/alexeycoder
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://localhost:8080/v1
    description: Test server
  - url: https://symcon.ru/v1
    description: Production server
...
paths:
  /users:
    post:
      tags:
        - users
      summary: Добавить пользователя
      description: Добавить пользователя
      operationId: addUser
      requestBody:
        description: Создание нового пользователя в базе
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '201':
          description: Created successfully
        '400':
          description: Invalid user data supplied
        '405':
          description: Invalid input
  /users/{userId}:
    get:
      tags:
        - users
      summary: Найти пользователя по идентификатору
      description: Получить одного пользователя
      operationId: findUserById
      parameters:
        - name: userId
          in: path
          description: Идентификатор пользователя
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
        '404':
          description: User not found

    put:
      tags:
        - users
      summary: Обновляет данные пользователя
      description: Обновляет данные пользователя по идентификатору
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          description: Идентификатор пользователя
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Обновление данных пользователя в базе
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        '200':
          description: OK
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
        '404':
          description: User not found

    delete:
      tags:
        - users
      summary: Удаляет пользователя
      description: Удаляет пользователя
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          description: Идентификатор пользователя
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Deleted successfully
        '400':
          description: Invalid ID supplied
        '401':
          description: Unauthorized
...
components:
  schemas:
    User:
      required:
        - username
        - password
        - name
      type: object
      properties:
        userId:
          type: integer
          format: int64
          examples: 
            - 123
        username:
          type: string
          format: email
          examples:
            - johnsmith@yandex.ru
          description: Адрес эл. почты пользователя в качестве имени для входа
        password:
          type: integer
          format: int64
          description: Хэш пароля
        name:
          type: string
          examples:
            - John Smith
          description: Отображаемое имя пользователя
...